Add files via upload
Leave Management System using PHP is a web-based application that enables organizations to manage employees' leave requests and approvals. It allows employees to submit leave requests online and administrators to review and approve them from a central dashboard. It also includes features such as leave tracking, reporting, and notifications. The system is designed to make leave management easier and more efficient for organizations. It provides real-time visibility into employee leave records, and allows managers to track the progress of leave requests. This system can be easily customized to meet the specific needs of an organization.